+++ tests/TestCase/Utility/AbstractBackupUtilityTest.php
@@ -131,7 +131,7 @@ class AbstractBackupUtilityTest extends TestCase
     /**
      * @return array<array{non-empty-string, non-empty-string}>
      */
-    public static function makeAbsoluteFilenameProvider(): array
+    public static function makeAbsolutePathProvider(): array
     {
         return [
             [
@@ -151,9 +151,9 @@ class AbstractBackupUtilityTest extends TestCase
 
     #[Test]
     #[DataProvider('makeAbsoluteFilenameProvider')]
-    public function testMakeAbsoluteFilename(string $expectedAbsolutePath, string $path): void
+    public function testMakeAbsolutePath(string $expectedAbsolutePath, string $path): void
     {
-        $result = $this->Utility->makeAbsoluteFilename($path);
+        $result = $this->Utility->makeAbsolutePath($path);
 
         $this->assertSame($expectedAbsolutePath, $result);
     }